+++ flywheel/__init__.py
@@ -1,6 +1,4 @@
 """ flywheel """
-__version__ = '0.4.0'
-
 from dynamo3 import (CheckFailed, ConditionalCheckFailedException, Binary,
                      STRING, NUMBER, BINARY, STRING_SET, NUMBER_SET,
                      BINARY_SET)
@@ -9,3 +7,5 @@ from .fields import Field, Composite, GlobalIndex, TypeDefinition, set_
 from .models import Model
 from .engine import Engine
 from .query import DuplicateEntityException, EntityNotFoundException
+
+__version__ = '0.4.1'